Various Stuff
- Lightbuffer is now packed to support HDR (well... MDR at least)
- Lights support HDR values now
- Fixed bug where spotlight shadows didn't get updated when the light
was rotated
- All lightshader were merged and put in default.fx
- Tweaked SSAO for a more realistic behaviour. Now only dark areas will
be affected by SSAO. If the user does not want this, he can tweak the
ssao to also be visible in bright areas
- Added additional scatter shader to bloom
- Added LOW, MEDIUM, HIGH and ULTRA quality setting for SSAO

+++ examples/fx/car.fx
@@ -297,6 +297,7 @@ PixelToFrame mainPS(vsOut In)
 	#else
 		PSOut.albedoAndEmissiveMask.xyz = SKIN_ALBEDO;
 	#endif
+	
 	//(team)color
 	#ifdef OBJECTCOLOR_A7
 		PSOut.albedoAndEmissiveMask.xyz += SKIN_COLOR * vecDiffuse;
@@ -325,7 +326,7 @@ PixelToFrame mainPS(vsOut In)
 	
 	//material data
 	PSOut.materialData.x = materialID; //material ID
-	PSOut.materialData.y = fPower/(float)255;//vecSkill17.z; //material Specular Power
+	PSOut.materialData.y = fPower/255.f;//vecSkill17.z; //material Specular Power
 	#ifdef GLOSSMAP
 		PSOut.materialData.z = SKIN_GLOSS; //material Specular Intensity
 	#else
@@ -341,6 +342,11 @@ PixelToFrame mainPS(vsOut In)
 	//PSOut.normalsAndDepth.w = 1;
 	//PSOut.albedoAndEmissiveMask.w = 1;
 	//PSOut.emissiveAndSpecular.w = 1;
+	
+	//Gamma Correction
+	//PSOut.albedoAndEmissiveMask.xyz = max(0, PSOut.albedoAndEmissiveMask.xyz-0.004);
+	//PSOut.albedoAndEmissiveMask.xyz = (PSOut.albedoAndEmissiveMask.xyz*(6.2*PSOut.albedoAndEmissiveMask.xyz+0.5))/(PSOut.albedoAndEmissiveMask.xyz*(6.2*PSOut.albedoAndEmissiveMask.xyz+1.7)+0.06);
+	//PSOut.albedoAndEmissiveMask.xyz = pow(PSOut.albedoAndEmissiveMask.xyz,2.2);
 	
 	return PSOut;
 }
